Reorganization of new assets/scripts
