+++ sdkproject/Assets/Mapbox/Examples/Drive/UI/LoadingPanelController.cs
@@ -1,26 +1,27 @@
-using Mapbox.Unity.MeshGeneration;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LoadingPanelController : MonoBehaviour
+namespace Mapbox.Examples
 {
-    public MapVisualizer MapVisualizer;
-    public GameObject Content;
+	using UnityEngine;
+	using Mapbox.Unity.Map;
 
-    void Awake()
-    {
-        MapVisualizer.OnMapVisualizerStateChanged += (s) =>
-        {
-            if (s == ModuleState.Finished)
-            {
-                Content.SetActive(false);
-            }
-            else if(s == ModuleState.Working)
-            {
-                Content.SetActive(true);
-            }
+	public class LoadingPanelController : MonoBehaviour
+	{
+		public MapVisualizer MapVisualizer;
+		public GameObject Content;
 
-        };
-    }
+		void Awake()
+		{
+			MapVisualizer.OnMapVisualizerStateChanged += (s) =>
+			{
+				if (s == ModuleState.Finished)
+				{
+					Content.SetActive(false);
+				}
+				else if (s == ModuleState.Working)
+				{
+					Content.SetActive(true);
+				}
+
+			};
+		}
+	}
 }
