+++ sdkproject/Assets/Mapbox/Examples/DynamicZoom/DynamicZoomMapVisualizer.asset
@@ -12,7 +12,6 @@ MonoBehaviour:
   m_Name: DynamicZoomMapVisualizer
   m_EditorClassIdentifier: 
   _factories:
-  - {fileID: 11400000, guid: 0db7771d70d7fc041a013fd2d957eec9, type: 2}
-  - {fileID: 11400000, guid: 5962f269ff9392341806bbe087abf85d, type: 2}
+  - {fileID: 0}
+  - {fileID: 0}
   _loadingTexture: {fileID: 2800000, guid: e2896a92727704803a9c422b043eae89, type: 3}
-  _loadingMaterial: {fileID: 2100000, guid: e486fb1a0b4233944af95fee011a22e0, type: 2}
