--- sdkproject/Assets/Mapbox/Examples/DynamicZoom/MapVisualization.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 9ea1dde608b65514eab316b799a45ce8
-folderAsset: yes
-timeCreated: 1503918086
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
