--- sdkproject/Assets/Mapbox/Examples/DynamicZoom/MapVisualization/DynamicZoomFlatTerrainFactory.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0a6523094c07d4ee683da5d826fce510, type: 3}
-  m_Name: DynamicZoomFlatTerrainFactory
-  m_EditorClassIdentifier: 
-  _baseMaterial: {fileID: 2100000, guid: b9f23e9bce724fa4daac57ecded470b8, type: 2}
-  _addCollider: 0
-  _addToLayer: 0
-  _layerId: 0
