--- sdkproject/Assets/Mapbox/Examples/DynamicZoom/MapVisualization/DynamicZoomFlatTerrainFactory.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 0db7771d70d7fc041a013fd2d957eec9
-timeCreated: 1504529848
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
