--- sdkproject/Assets/Mapbox/Examples/DynamicZoom/MapVisualization/DynamicZoomMapImageFactory.asset
@@ -1,23 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2b10536479dade041b6db9893fdf723a, type: 3}
-  m_Name: DynamicZoomMapImageFactory
-  m_EditorClassIdentifier: 
-  _mapIdType: 1
-  _customStyle:
-    Name: 
-    Id: mapbox://styles/mapbox/satellite-v9
-    Modified: 
-    UserName: 
-  _mapId: mapbox://styles/mapbox/satellite-v9
-  _useCompression: 0
-  _useMipMap: 1
-  _useRetina: 0
