--- sdkproject/Assets/Mapbox/Examples/DynamicZoom/MapVisualization/DynamicZoomMapImageFactory.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5962f269ff9392341806bbe087abf85d
-timeCreated: 1504614847
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
