+++ sdkproject/Assets/Mapbox/Examples/Scripts/DynamicZoomCameraMovement.cs.meta
